% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoperationSHP.R
\name{.summary_ts}
\alias{.summary_ts}
\title{TS Summary}
\usage{
.summary_ts(ts_data)
}
\arguments{
\item{ts_data}{time series from polygon}
}
\value{
Data set with summary data
}
\description{
Acquire summary of time series tibble from shapefile
}
\author{
Adeline Maciel
}
