% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSfilter.R
\name{TSfilter}
\alias{TSfilter}
\title{TS Filter}
\usage{
TSfilter(ts_data = NULL,
type_filter = c("No-filter", "Whittaker", "Savitsky-Golay"),
wh_lambda = "1.0", wh_differences = "3", sg_order = "3",
sg_scale = "1")
}
\arguments{
\item{ts_data}{data set with point via WTSS}

\item{type_filter}{name of the filter to be applied}

\item{wh_lambda}{smoothing factor to be applied (default 1.0). See sits_whittaker in sits package {sits}}

\item{wh_differences}{the order of differences of contiguous elements (default 3). See sits_whittaker in sits package {sits}}

\item{sg_order}{filter order (default 3). See sits_sgolay in sits package {sits}}

\item{sg_scale}{time scaling (default 1). See sits_sgolay in sits package {sits}}
}
\value{
Plot time series filtered
}
\description{
Apply a filter over time series based on SITS package \url{https://github.com/e-sensing/sits}
}
\author{
Adeline Maciel
}
