% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoperationSHP.R
\name{TSoperationSHP}
\alias{TSoperationSHP}
\title{TS Operation shapefile}
\usage{
TSoperationSHP(name_service = c("WTSS-INPE", "SATVEG"),
coverage = c("MOD13Q1", "terra", "aqua", "comb"),
bands = c("ndvi", "evi", "nir", "mir", "blue", "red"),
start_date = NULL, end_date = NULL, pre_filter = "1", geojson_points = NULL)
}
\arguments{
\item{name_service}{information of service, like WTSS-INPE or SATVEG.}

\item{coverage}{name of coverage from service.}

\item{bands}{set of bands from coverage}

\item{start_date}{first date of interval}

\item{end_date}{last date of the interval}

\item{pre_filter}{a string ("0" none, "1" no data correction, "2" cloud correction, "3" no data and cloud correction). Information of sits package.}

\item{geojson_points}{receive a GeoJSON with all points from polygon}
}
\value{
Data set with time series data
}
\description{
Acquire time series tibble from service
}
\author{
Adeline Maciel
}
