% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSplot.R
\name{TSplot}
\alias{TSplot}
\title{TS Plot}
\usage{
TSplot(ts_data = NULL)
}
\arguments{
\item{ts_data}{data set with point via WTSS.}
}
\value{
Plot time series information according to the bands and time
}
\description{
Plot acquired time series from service
}
\author{
Adeline Maciel
}
