% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoperation.R
\name{TSoperation}
\alias{TSoperation}
\title{TS Operation}
\usage{
TSoperation(name_service = c("WTSS", "SATVEG"),
coverage = c("MOD13Q1", "terra", "aqua", "comb"), longitude = NULL,
latitude = NULL, bands = c("ndvi", "evi", "nir", "mir", "blue", "red"),
start_date = NULL, end_date = NULL)
}
\arguments{
\item{name_service}{information of service, like WTSS or SATVEG.}

\item{coverage}{name of coverage from service.}

\item{longitude}{longitude information.}

\item{latitude}{latitude information.}

\item{bands}{set of bands from coverage}

\item{start_date}{first date of interval}

\item{end_date}{last date of the interval}
}
\value{
Data set with time series data
}
\description{
Acquire time series tibble from service
}
\author{
Adeline Maciel
}
